terminal: Implement enabling raw mode on terminal
By default, the terminal starts in canonical mode - also called cooked
mode. In this mode, keyboard input is only sent to programs when the
user presses Enter.

However, in a text editor, we want to process each keypress as it
comes in. This mode is called the raw mode and is achieved by
turning off many flags in the terminal.

Start by disabling echoing in the terminal.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>