terminal: Turn off all output post processing
By default, the termonal also translates each newline ("\n") we print
into a carriage return followed by a newline ("\r\n"). The terminal
requires both of these characters in order to start a new line of text.

The carriage return moves the cursor back to the beginning of the
current line, and the newline moves the cursor down a line,
scrolling the screen if necessary

Disable this, and all other, output post processing.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>