main: Add carriage returns when printing newline
After disabling output post processing, we see that the newline
characters we’re printing are only moving the cursor down, and not to
the left side of the screen.

Add carriage returns to our printf() statements. From now on, we’ll
have to write out the full "\r\n" whenever we want to start a new line.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>