editor: Refresh the editor's UI after each keypress
We use the VT100 escape sequences which are supported very widely by
modern terminal emulators.

Escape sequences begin with the escape character - \x1b, or 27 in
decimal. We write two escape sequences to refresh the screen :

The first is the ED – Erase In Display sequence, which erases some or
all of the characters in the display according to the parameter. We use
the argument 2, which says to clear the entire screen.

The second is the CUP – Cursor Position sequence, which moves the active
position to the position specified by the parameters. The default
condition with no parameters present is equivalent to a cursor to home
action, or 1,1 position.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>