editor: Add fallback for getting terminal window size
ioctl() isn’t guaranteed to be able to request the window size on all
systems, so add a fallback method.

The strategy is to position the cursor at the bottom-right of the screen,
then use escape sequences that let us query the position of the cursor.

We send two escape sequences, the C command (Cursor Forward) and the B
command (Cursor Down) passing 999 as arguments.

Next, we use the N command (Device Status Report) with argument 6,
which returns the active cursor position. The reply is an escape
sequence documented as Cursor Position Report.

The format is : \x1b[13;194
where 13 and 194 can be any two numbers.

These numbers are used as the numbers of rows and columns.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>