cilo: Implement struct AppendBuffer
It’s not a good idea to make a whole bunch of small write()’s every
time we refresh the screen. It would be better to do one big write(),
to make sure the whole screen updates at once. Otherwise there could
be small unpredictable pauses between write()’s, which would cause an
annoying flicker effect. Not to mention the performance cost.

Create our own dynamic string type AppendBuffer which supports one
operation - insertion.

We'll use this to replace all write() calls in editor, and then
flush the buffer as the last step of refreshing the display.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>