append_buffer: Fix to actually use our preallocated buffer
We were always reallocating the buffer on every insert, disregarding
our initial preallocation.

Instead, only realloc() if we exceed the initial size when growing.
Otherwise, modify and memcpy() to the end of ab->buffer itself.

Let's also be pedantic and check ab->buffer for NULL since we passed
that responsibility from buffer_init() to buffer_insert().

Increase BUFFER_INITIAL_SIZE to 256 as well. This is not a complete
solution though. Will make AppendBuffer an actual dynamic array with
length and capacity variables and grow with 2x factor. But this is
good enough for now.

This is what a good shower does!

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>