editor: Map arrow keys to wasd
Pressing an arrow key sends multiple bytes as input to our program.
These bytes are in the form of an escape sequence that starts with
'\x1b', '[', followed by an 'A', 'B', 'C', or 'D' depending on which
of the four arrow keys was pressed.

If we read an escape character, we immediately read two more bytes
into the seq buffer. If either of these reads time out
(after 0.1 seconds), then we assume the user just pressed the Escape
key and return that.

Otherwise we look to see if the escape sequence is an arrow key escape
sequence. If it is, we just return the corresponding wasd character,
for now.

If it’s not an escape sequence we recognize, we just return the escape
character.

This allows moving the cursor using the arrow keys, but still leaves
wasd mapped to updating the cursor as well. Have to fix that.

Add a missimg break after move_cursor() too.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>