editor: Add support for Home and End keys
There are many different escape sequences that could be sent by these
keys, depending on the OS, or the terminal emulator used.

The Home key could be sent as <esc>[1~, <esc>[7~, <esc>[H, or <esc>OH.
The End key could be sent as <esc>[4~, <esc>[8~, <esc>[F, or <esc>OF.

Handle all of these and move cursor to the left or right edges of the
screen when pressed.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>