cilo: Add support for opening files
We store an array of EditorRow in the EditorState struct to store
multiple lines. We dynamically grow this array as required - with
an initial capacity of 16, doubling it when required.

file_io.c handles opening, reading and calling functions to store lines.
Pass a NULL ptr to getline() call so that getline handles all the memory
for us.

We strip off the newline or carriage return at the end of read lines
before storing them. We know each EditorRow represents one line/row of
text, so there’s no use storing a newline character.

When displaying contents of a file - truncate if the length of a line is
more than the size of the terminal, for now.

Also add a check to only display the welcome message when the user starts
the editor with no arguments, i.e., when not opening a file.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>