editor: Implement vertical scrolling
The row_offset member keeps track of the currently scrolled to line
from the top of the file.

We check if the cursor has moved outside of the visible window, and if
so, adjust editor.row_offset so that the cursor is just inside the
visible window. This is done in update_scroll().

The first if statement checks if the cursor is above the visible window,
and if so, scrolls up to where the cursor is. The second if statement
checks if the cursor is past the bottom of the visible window.

Also allow the user to scroll past the end of the screen, but not past
the end of the file.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>