editor: Implement horizontal scrolling
Follows much of the same logic as vertical scrolling.

To display each row at the column offset, we’ll use col_offset as an
index into the char* line member of each EditorRow we display, and
subtract the number of characters that are to the left of the offset
from the length of the row. The result of this subtraction can be a
negative number, i.e., the user scrolled past the end of a line - so
clamp this to 0.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>