editor/input: Limit scrolling to the right
Both cursor_x and cursor_y refer to the cursor’s position within the
file, not its position on the screen. Let's start limiting the values
of these members to only ever point to valid positions in the file.

Start by not allowing the user to scroll past the end of the current
line. Only move cursor right if cursor_x is to the left of the end of
the current line.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>