editor/state: Refactor displaying file
Let's factor out the subtraction to a const variable, which can be
negative meaning the user scrolled horizontally past the end of the
line. In that case, set length to 0 so that nothing is displayed on
the line.

This generates better assembly on gcc at O3, whereas clang shows no
difference.

Remove unnecessary grow_rows_array() as well. Too verbose.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>