editor/state: Implement rendering tabs
First, we count the number of tabs to know how much memory
to allocate for render_chars.

The maximum number of characters needed for each tab is CILO_TAB_STOP.
row->line_length already counts 1 for each tab, so we multiply the number
of tabs by (CILO_TAB_STOP - 1) and add that to row->line_length to get
the maximum amount of memory we’ll need for the rendered row.

Loop over the line_chars and keep appending spaces for each encountered
tab until we get to a tab stop, which is a column divisible by CILO_TAB_STOP.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>