editor/state: Handle tabs and cursor interaction
Before this change, when we positioned the cursor on the screen, we were
still assuming each character takes up only one column on the screen.

To fix this, introduce a new horizontal co-ordinate member in our
EditorState struct which is an index into the render_chars field.

If there are no tabs on the current line, then editor.render_x will be
the same as editor.cursor_x. If there are tabs, then editor.render_x
will be greater by however many extra spaces those tabs take up when
rendered.

For each character, if it’s a tab we use render_x % CILO_TAB_STOP to
find out how many columns we are to the right of the last tab stop,
and then subtract that from CILO_TAB_STOP - 1 to find out how many
columns we are to the left of the next tab stop.

We add that amount to render_x to get just to the left of the next tab
stop, and then the unconditional render_x++ statement gets us right on
the next tab stop.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>