editor/row: Implement er_insert_character()
Inserts a character at a given index in a EditorRow, and updates
the render fields accordingly.

Allocate one extra byte for the line_chars of the row (we add 2
because we also have to make room for the null byte), and use
memmove() to make room for the new character. We increment line_length,
and then actually assign the character to its position in the array.

Finally, call er_update_render() so that the render fields get updated
with the new row content.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>