editor: Implement inserting characters
cilo is on its way to becoming a text editor, and not just a text viewer.

We create a new module called editor/operations.h.
This module would contain functions we'll call from editor/input, when
we're mapping keypresses to various text editing operations.

editor/operations does not have to worry about the details of modifying
a struct EditorRow, while editor/row does not have to worry about where
the cursor is.

Map all unmapped keys to editor_op_insert_char() by adding it to the
default case.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>