file_io: Implement saving file to disk
er_to_string() converts of array of EditorRow structs, stored in the
struct EditorState editor, into a single string that is ready to be
written out to a file.

If our filename is NULL, return for now.

Otherwise, we call er_to_string() and write() the resulting string to
the file path. We tell open() we want to create a new file if it doesn’t
already exist (O_CREAT), and we want to open it for reading and writing
(O_RDWR), passing the mode argument 0644.

The normal way to overwrite a file is to pass the O_TRUNC flag to open(),
which truncates the file completely, making it an empty file, before
writing the new data into it. By truncating the file ourselves to the same
length as the data we are planning to write into it, we are making the
whole overwriting operation a little bit safer in case the ftruncate()
call succeeds but the write() call fails.

In this case, the file would still contain most of the data it had before.
But if the file was truncated completely by the open() call and then the
write() failed, and all data is lost.

open() and ftruncate() both return -1 on error. We expect write() to
return the number of bytes we told it to write. Whether or not an error
occurred, we ensure that the file is closed and the memory that buf
points to is freed.

Map this file_io function to the Ctrl+S key and update our startup
help message.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>