editor: Implement deleting characters
Similar to how insertion was implemented, except we don't need to
manage memory in er_delete_character(). We just use memmove() to
overwrite the deleted character with the characters that come after it.

In editor/operations, if the cursor’s past the end of the file, we return
immediately. Otherwise, we get the EditorRow the cursor is on, and if
there is a character to the left of the cursor, we delete it and move
the cursor one to the left.

Delete key is implemented as pressing ArrowRight followed by a backspace.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>