editor: Implement inserting new line by pressing Enter
If we’re at the beginning of a line, all we have to do is insert a new
blank row before the line we’re on.

Otherwise, we have to split the line we’re on into two rows.
First we call editor_state_insert_line() and pass it the characters on
the current row that are to the right of the cursor. That creates a new
row after the current one, with the correct contents. Then we reassign
the row pointer, because editor_state_insert_line() calls realloc(),
which might move memory around on us and invalidate the pointer.

Then we truncate the current row’s contents by setting its size to the
position of the cursor, and we call er_update_render() on the truncated
row.

In both cases, we increment editor.cursor_y, and set editor.cursor_x to
0 to move the cursor to the beginning of the row.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>