editor/input: Implement function to take input when prompted
Pretty straightforward.

The only thing to note is that the passed in const char* prompt is
expected to be a format string containing a %s, which is where the
user's input will be displayed.

We also check if the input key isn't one of the special keys in our
EditorKey enum, which have high int values - by making sure it is less
than 128.

When an input prompt is cancelled by pressing Esc, return NULL.
This should be handled by all callers.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>