editor/operations: Implement search operation
We loop through all the rows of the file and use strstr() to check if
a given query is a substring of the current row.

Update cursor_x by subtracting the row->render_chars pointer from the
matched pointer, since match is a pointer to the same render_chars string.

A helper method editor_rx_to_cx() is used to convert the index into the
render_chars string to an index in the chars_string, by accounting for tabs,
before assigning it to cursor_x.

We set row_offset to the bottom of the file, so that when update_scroll()
is called for the next redraw - the matching line will be at the very top
of the screen.

The helper function works as follows :
We loop through the chars string, calculating the current rx value
(curr_rx) as we go. We stop when curr_rx hits the given rx value and
return cx.

The return statement at the very end is just in case the caller provided
an rx that’s out of range, which shouldn’t happen. :)

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>