editor/operations: Pass callback to prompt for incremental search
op_search_callback() is our callback function when asking the user for
input when searching for a query.

In the callback, we check if the user pressed Enter or Escape, in which
case they are leaving search mode so we return immediately instead of
doing another search.

Otherwise, after any other keypress, we do another search for the current
query string.

Cool. :)

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>