editor/operations: Use arrow keys to search forward/backward
We use two static variables to specify the index of last matched row
(or -1 if there was no last match), and direction of the search.

We always reset last_match_idx to -1 unless an arrow key was pressed.
We’ll only advance to the next or previous match when an arrow key
is pressed.

Also always set direction to 1 unless the left or up arrow key was
pressed. We always search in the forward direction unless the user
specifically asks to search backwards from the last match.

current_idx is the index of the current row we are searching. If there
was a last match, it starts on the line after (or before, if we’re
searching backwards). If there wasn’t a last match, it starts at the top
of the file and searches in the forward direction to find the first match.

Wrap around current to go from end of the file to the beginning of the
file, and vice versa.

When we find a match, we set last_match_idx to current_idx; when the user
presses the arrow keys, we’ll start the next search from that point.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>