editor/row: Add highlight member to each row
Prep for syntax highlighting.

highlight is an array of bytes, where each value will correspond to a
character in render_chars, and will tell us the information required
for syntax highlighting for it, eg - if it's part of a string, a
comment etc.

This is all the info we need to implement syntax highlighting
in cilo.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>