editor/row: Fix crash when highlighting empty line
When render_length is 0, i.e., the last character is deleted on a line
and er_update_highlight() is called, the realloc call frees row->highlight
and we hit the die() path.

Since, er_update_highlight() is called from er_update_render() in this
case - we can simply check if er_update_render() was called on a row
with length 0, in which case render_length is set to 0.

Could also have added an implicit +1 in the realloc size parameter to
avoid branches.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>